# infra/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anonymizer API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://{host}{basePath}",
      "variables": {
        "basePath": {
          "enum": [
            "anonymizer/v1"
          ],
          "default": "/anonymizer/v1"
        },
        "host": {
          "enum": [
            "https://api.dev.platform.pagopa.it",
            "https://api.uat.platform.pagopa.it",
            "https://api.platform.pagopa.it"
          ],
          "default": "https://api.platform.pagopa.it"
        }
      }
    }
  ],
  "paths": {
    "/info": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Get application info",
        "description": "Liveness & readiness endpoint. Returns application name, version, and environment.",
        "operationId": "info_info_get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InfoResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/anonymize": {
      "post": {
        "tags": [
          "Anonymize"
        ],
        "summary": "Anonymize text",
        "description": "Anonymizes the provided text using Presidio.",
        "operationId": "anonymize_endpoint_anonymize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonymizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnonymizeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "InfoResponse": {
        "title": "InfoResponse",
        "required": [
          "name",
          "version",
          "environment"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "title": "ErrorResponse",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        }
      },
      "AnonymizeResponse": {
        "title": "AnonymizeResponse",
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "title": "Text",
            "type": "string",
            "description": "Anonymized text"
          }
        }
      },
      "AnonymizeRequest": {
        "title": "AnonymizeRequest",
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "title": "Text",
            "type": "string",
            "description": "Text to be anonymized"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  },
  "tags": [
    {
      "name": "Info",
      "description": "Liveness & readiness endpoints"
    },
    {
      "name": "Anonymize",
      "description": "Text anonymization endpoints"
    }
  ]
}